Create FreqArray-for Multiple Search
This is the code for creating a frequency array for optimized  multiple search  with single traversal.
First we took an array as an input. Than we created a multidimensional array named freqarray. It stores flags and index of occurance in original array for array elements corresposnding to each of them in the frequency array. And now we directly search from freqarray  and get our answer without travesing the array again and again.