Create SecondLargest Element in Array Method 2
In this code we are finding second largest element in a user input array. First we find the largest element using max method then we change the max element to Integer.MIN_VALUE and again find largest element using max method, which will be second largest element of the original array.